requirements.txt
deepgram-sdk
